feat: 관리자 기능 추가

+++ src/main/java/environment/project/dto/ClubInfoDTO.java
@@ -7,7 +7,7 @@ import java.util.Date;
 @Data
 public class ClubInfoDTO {
     private Long idx;
-    private Long clubUserGrade;
+    private int clubUserGrade;
     private boolean clubUsed;
     private Date clubJoinDate;
     private Date clubApproveDate;
